fix that was preventing startup on brand new nodes.

+++ atlas-xcache/dark-data-cleaner.py
@@ -59,6 +59,8 @@ for i in range(3):
     dirs_deleted = 0
     dt = os.walk(BASE_DIR)
     for d in dt:
+        if d == BASE_DIR:
+            continue
         if not d[1] and not d[2]:
             os.rmdir(d[0])
             dirs_deleted += 1
